+++ src/test/java/com/fastcampus/projectboard/domain/repository/JpaRepositoryTest.java
@@ -1,6 +1,8 @@
 package com.fastcampus.projectboard.domain.repository;
 
 import com.fastcampus.projectboard.domain.Article;
+import com.fastcampus.projectboard.repository.ArticleCommentRepository;
+import com.fastcampus.projectboard.repository.ArticleRepository;
 import config.JpaConfig;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
